Edit Wedding functionality
Implemented the edit wedding info function
--Currently placed in the tool strip in the Main Window

There are some issues with the way it is currently set up - initial
dialog that I made for wedding creation has two text boxes for entry -
wedding name and wedding date.  Editing an existing wedding makes this a
bit more difficult as there is only one column in the table; parsing the
wed_party_name to split the name/date again is not reliably accurate.
We could add another column to wed_party for wed_party_date if we really
want to push it.

I have it set up to just populate the top textbox at the moment so it
works.  I've run into a fair number of bugs/thoughts while putting this
together that we can go over when we meet.