Initial edit wedding implementation
Basics of Edit Wedding are implemented in the new version
TODO: current implementation loads the full string into wedding party
name, still need to write the parse to split the string and populate the
combo boxes.
-Update is pushed straight to the database via SQLiteCommand rather than
the data table

BUGS: Noticed this on LoadForm as well - if entry fields are blank the
dialog pops up the message box asking for data entry - click ok/close
the dialog and it kills the whole form rather than focusing back on the
needed data.  Didn't have a chance to run through this this morning.